Rename to node-fetch-polyfill
Upstream [is not keen on incorporating breaking
changes](https://github.com/bitinn/node-fetch/pull/140) like the removal of
charset detection & the introduction of ReadableStream. In the meantime, we
need a spec conforming `fetch` compatible polyfill that can be used in
ServiceWorker code. So, for now, the path of least resistance seems to be to
rename the project to node-fetch-polyfill & bump the version to 2.0.0 to
signal breaking API changes.